Add escape hatch for games on a webhost with no slot files

+++ ap-clean.sh
@@ -4,8 +4,10 @@ if [[ -n $AP_PLAYER && $AP_ROOM =~ ^(https?://archipelago.gg/room/)?([^/]+)$ ]];
     curl "${AP_ROOM_BASE:-https://archipelago.gg}/api/room_status/${BASH_REMATCH[2]}" -o /tmp/room
     mkdir /factorio/mods
     cd /factorio/mods
-    rm AP-*.zip
-    curl -OJ "${AP_ROOM_BASE:-https://archipelago.gg}$(jq -r '.downloads[.players | map(.[0] == "'"$AP_PLAYER"'") | index(true)].download' /tmp/room)"
+    if [[ -z $AP_NO_MANAGE_MOD ]]; then
+        rm AP-*.zip
+        curl -OJ "${AP_ROOM_BASE:-https://archipelago.gg}$(jq -r '.downloads[.players | map(.[0] == "'"$AP_PLAYER"'") | index(true)].download' /tmp/room)"
+    fi
     AP_URL="${AP_PROTOCOL:-wss}://$AP_PLAYER:$AP_PASSWORD@${AP_URL_BASE:-archipelago.gg:}$(jq -r '.last_port' /tmp/room)";
 fi
